# ENH: Improve load time of displays that re-use the same UI file repeatedly