@ZLLentz 2023-02-03
I should have read the full PR test- I went to test pmps-ui just as you did, and I observed the same speedup. The speedup was the same (6-7 seconds) in the full production environment. That's a lot of time saved that was previously spent recompiling the same few files thousands of times.

This is pretty clever. I didn't connect in my head that the uic compilation has an intermediate step that generates python code that can be subbed into with the same templating mechanisms used on the file load.

This should help a lot for screens with template repeaters and similar constructs.